requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
